+++ src/list_files/get_paths.rs
@@ -12,10 +12,19 @@ pub fn main(
 }
 
 fn get_path(base_folder: &path::Path, path: path::PathBuf) -> model::Result<model::Path> {
-    let relative = get_path_string(path.strip_prefix(base_folder)?)?;
+    let relative = get_path_components(path.strip_prefix(base_folder)?)?;
     let absolute = get_path_string(&path)?;
 
     Ok(model::Path { relative, absolute })
+}
+
+fn get_path_components(path: &path::Path) -> model::Result<Vec<String>> {
+    path.iter()
+        .map(|component| match component.to_str() {
+            None => Err(model::Error::PathInvalidUnicode(path.to_path_buf())),
+            Some(string) => Ok(String::from(string)),
+        })
+        .collect()
 }
 
 fn get_path_string(path: &path::Path) -> model::Result<String> {
@@ -42,11 +51,11 @@ mod tests {
         let actual = actual.unwrap();
         let expected = vec![
             model::Path {
-                relative: String::from('c'),
+                relative: vec![String::from('c')],
                 absolute: String::from("/a/b/c"),
             },
             model::Path {
-                relative: String::from("a/b"),
+                relative: vec![String::from('a'), String::from('b')],
                 absolute: String::from("/a/b/a/b"),
             },
         ];